[AS7-4090] Move subsystem describe operations to the correct class.
