+++ clustering/infinispan/src/main/java/org/jboss/as/clustering/infinispan/subsystem/InfinispanSubsystemDescribe.java
@@ -22,6 +22,13 @@
 
 package org.jboss.as.clustering.infinispan.subsystem;
 
+import static org.jboss.as.clustering.infinispan.subsystem.BaseJDBCStoreResource.COMMON_JDBC_STORE_ATTRIBUTES;
+import static org.jboss.as.clustering.infinispan.subsystem.BaseStoreResource.COMMON_STORE_ATTRIBUTES;
+import static org.jboss.as.clustering.infinispan.subsystem.BinaryKeyedJDBCStoreResource.BINARY_KEYED_JDBC_STORE_ATTRIBUTES;
+import static org.jboss.as.clustering.infinispan.subsystem.MixedKeyedJDBCStoreResource.MIXED_KEYED_JDBC_STORE_ATTRIBUTES;
+import static org.jboss.as.clustering.infinispan.subsystem.StringKeyedJDBCStoreResource.STRING_KEYED_JDBC_STORE_ATTRIBUTES;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.ADD;
+
 import org.jboss.as.controller.AttributeDefinition;
 import org.jboss.as.controller.OperationContext;
 import org.jboss.as.controller.OperationDefinition;
@@ -30,6 +37,7 @@ import org.jboss.as.controller.OperationStepHandler;
 import org.jboss.as.controller.PathAddress;
 import org.jboss.as.controller.SimpleOperationDefinitionBuilder;
 import org.jboss.as.controller.descriptions.ModelDescriptionConstants;
+import org.jboss.as.controller.operations.common.Util;
 import org.jboss.as.controller.registry.Resource;
 import org.jboss.dmr.ModelNode;
 import org.jboss.dmr.ModelType;
@@ -180,28 +188,28 @@ public class InfinispanSubsystemDescribe implements OperationStepHandler {
             ModelNode locking = cache.getValue().get(ModelKeys.LOCKING, ModelKeys.LOCKING_NAME);
             ModelNode lockingAddress = address.clone();
             lockingAddress.add(ModelKeys.LOCKING, ModelKeys.LOCKING_NAME);
-            result.add(CacheConfigOperationHandlers.createOperation(LockingResource.LOCKING_ATTRIBUTES, lockingAddress, locking));
+            result.add(createOperation(LockingResource.LOCKING_ATTRIBUTES, lockingAddress, locking));
         }
         // command to recreate the transaction configuration
         if (cache.getValue().get(ModelKeys.TRANSACTION, ModelKeys.TRANSACTION_NAME).isDefined()) {
             ModelNode transaction = cache.getValue().get(ModelKeys.TRANSACTION, ModelKeys.TRANSACTION_NAME);
             ModelNode transactionAddress = address.clone();
             transactionAddress.add(ModelKeys.TRANSACTION, ModelKeys.TRANSACTION_NAME);
-            result.add(CacheConfigOperationHandlers.createOperation(TransactionResource.TRANSACTION_ATTRIBUTES, transactionAddress, transaction));
+            result.add(createOperation(TransactionResource.TRANSACTION_ATTRIBUTES, transactionAddress, transaction));
         }
         // command to recreate the eviction configuration
         if (cache.getValue().get(ModelKeys.EVICTION, ModelKeys.EVICTION_NAME).isDefined()) {
             ModelNode eviction = cache.getValue().get(ModelKeys.EVICTION, ModelKeys.EVICTION_NAME);
             ModelNode evictionAddress = address.clone();
             evictionAddress.add(ModelKeys.EVICTION, ModelKeys.EVICTION_NAME);
-            result.add(CacheConfigOperationHandlers.createOperation(EvictionResource.EVICTION_ATTRIBUTES, evictionAddress, eviction));
+            result.add(createOperation(EvictionResource.EVICTION_ATTRIBUTES, evictionAddress, eviction));
         }
         // command to recreate the expiration configuration
         if (cache.getValue().get(ModelKeys.EXPIRATION, ModelKeys.EXPIRATION_NAME).isDefined()) {
             ModelNode expiration = cache.getValue().get(ModelKeys.EXPIRATION, ModelKeys.EXPIRATION_NAME);
             ModelNode expirationAddress = address.clone();
             expirationAddress.add(ModelKeys.EXPIRATION, ModelKeys.EXPIRATION_NAME);
-            result.add(CacheConfigOperationHandlers.createOperation(ExpirationResource.EXPIRATION_ATTRIBUTES, expirationAddress, expiration));
+            result.add(createOperation(ExpirationResource.EXPIRATION_ATTRIBUTES, expirationAddress, expiration));
         }
 
         // command to recreate the cache store configuration
@@ -210,7 +218,7 @@ public class InfinispanSubsystemDescribe implements OperationStepHandler {
             ModelNode store = cache.getValue().get(ModelKeys.STORE, ModelKeys.STORE_NAME);
             ModelNode storeAddress = address.clone();
             storeAddress.add(ModelKeys.STORE, ModelKeys.STORE_NAME);
-            result.add(CacheConfigOperationHandlers.createStoreOperation(BaseStoreResource.COMMON_STORE_ATTRIBUTES, storeAddress, store,
+            result.add(createStoreOperation(BaseStoreResource.COMMON_STORE_ATTRIBUTES, storeAddress, store,
                     StoreResource.STORE_ATTRIBUTES));
             addStoreWriteBehindConfigCommands(store, storeAddress, result);
             addCacheStorePropertyCommands(store, storeAddress, result);
@@ -219,7 +227,7 @@ public class InfinispanSubsystemDescribe implements OperationStepHandler {
             ModelNode store = cache.getValue().get(ModelKeys.FILE_STORE, ModelKeys.FILE_STORE_NAME);
             ModelNode storeAddress = address.clone();
             storeAddress.add(ModelKeys.FILE_STORE, ModelKeys.FILE_STORE_NAME);
-            result.add(CacheConfigOperationHandlers.createStoreOperation(BaseStoreResource.COMMON_STORE_ATTRIBUTES, storeAddress, store,
+            result.add(createStoreOperation(BaseStoreResource.COMMON_STORE_ATTRIBUTES, storeAddress, store,
                     FileStoreResource.FILE_STORE_ATTRIBUTES));
             addStoreWriteBehindConfigCommands(store, storeAddress, result);
             addCacheStorePropertyCommands(store, storeAddress, result);
@@ -228,7 +236,7 @@ public class InfinispanSubsystemDescribe implements OperationStepHandler {
             ModelNode store = cache.getValue().get(ModelKeys.STRING_KEYED_JDBC_STORE, ModelKeys.STRING_KEYED_JDBC_STORE_NAME);
             ModelNode storeAddress = address.clone();
             storeAddress.add(ModelKeys.STRING_KEYED_JDBC_STORE, ModelKeys.STRING_KEYED_JDBC_STORE_NAME);
-            result.add(CacheConfigOperationHandlers.createStringKeyedStoreOperation(storeAddress, store));
+            result.add(createStringKeyedStoreOperation(storeAddress, store));
             addStoreWriteBehindConfigCommands(store, storeAddress, result);
             addCacheStorePropertyCommands(store, storeAddress, result);
         }
@@ -236,7 +244,7 @@ public class InfinispanSubsystemDescribe implements OperationStepHandler {
             ModelNode store = cache.getValue().get(ModelKeys.BINARY_KEYED_JDBC_STORE, ModelKeys.BINARY_KEYED_JDBC_STORE_NAME);
             ModelNode storeAddress = address.clone();
             storeAddress.add(ModelKeys.BINARY_KEYED_JDBC_STORE, ModelKeys.BINARY_KEYED_JDBC_STORE_NAME);
-            result.add(CacheConfigOperationHandlers.createBinaryKeyedStoreOperation(storeAddress, store));
+            result.add(createBinaryKeyedStoreOperation(storeAddress, store));
             addStoreWriteBehindConfigCommands(store, storeAddress, result);
             addCacheStorePropertyCommands(store, storeAddress, result);
         }
@@ -244,7 +252,7 @@ public class InfinispanSubsystemDescribe implements OperationStepHandler {
             ModelNode store = cache.getValue().get(ModelKeys.MIXED_KEYED_JDBC_STORE, ModelKeys.MIXED_KEYED_JDBC_STORE_NAME);
             ModelNode storeAddress = address.clone();
             storeAddress.add(ModelKeys.MIXED_KEYED_JDBC_STORE, ModelKeys.MIXED_KEYED_JDBC_STORE_NAME);
-            result.add(CacheConfigOperationHandlers.createMixedKeyedStoreOperation(storeAddress, store));
+            result.add(createMixedKeyedStoreOperation(storeAddress, store));
             addStoreWriteBehindConfigCommands(store, storeAddress, result);
             addCacheStorePropertyCommands(store, storeAddress, result);
         }
@@ -252,7 +260,7 @@ public class InfinispanSubsystemDescribe implements OperationStepHandler {
             ModelNode store = cache.getValue().get(ModelKeys.REMOTE_STORE, ModelKeys.REMOTE_STORE_NAME);
             ModelNode storeAddress = address.clone();
             storeAddress.add(ModelKeys.REMOTE_STORE, ModelKeys.REMOTE_STORE_NAME);
-            result.add(CacheConfigOperationHandlers.createStoreOperation(BaseStoreResource.COMMON_STORE_ATTRIBUTES, storeAddress, store,
+            result.add(createStoreOperation(BaseStoreResource.COMMON_STORE_ATTRIBUTES, storeAddress, store,
                     RemoteStoreResource.REMOTE_STORE_ATTRIBUTES));
             addStoreWriteBehindConfigCommands(store, storeAddress, result);
             addCacheStorePropertyCommands(store, storeAddress, result);
@@ -274,7 +282,7 @@ public class InfinispanSubsystemDescribe implements OperationStepHandler {
             ModelNode stateTransfer = cache.getValue().get(ModelKeys.STATE_TRANSFER, ModelKeys.STATE_TRANSFER_NAME);
             ModelNode stateTransferAddress = address.clone();
             stateTransferAddress.add(ModelKeys.STATE_TRANSFER, ModelKeys.STATE_TRANSFER_NAME);
-            ModelNode createOperation = CacheConfigOperationHandlers.createOperation(StateTransferResource.STATE_TRANSFER_ATTRIBUTES, stateTransferAddress, stateTransfer);
+            ModelNode createOperation = createOperation(StateTransferResource.STATE_TRANSFER_ATTRIBUTES, stateTransferAddress, stateTransfer);
             result.add(createOperation);
         }
     }
@@ -285,7 +293,7 @@ public class InfinispanSubsystemDescribe implements OperationStepHandler {
             ModelNode writeBehind = store.get(ModelKeys.WRITE_BEHIND, ModelKeys.WRITE_BEHIND_NAME);
             ModelNode writeBehindAddress = address.clone();
             writeBehindAddress.add(ModelKeys.WRITE_BEHIND, ModelKeys.WRITE_BEHIND_NAME);
-            ModelNode createOperation = CacheConfigOperationHandlers.createOperation(StoreWriteBehindResource.WRITE_BEHIND_ATTRIBUTES, writeBehindAddress, writeBehind);
+            ModelNode createOperation = createOperation(StoreWriteBehindResource.WRITE_BEHIND_ATTRIBUTES, writeBehindAddress, writeBehind);
             result.add(createOperation);
         }
     }
@@ -296,40 +304,71 @@ public class InfinispanSubsystemDescribe implements OperationStepHandler {
              for (Property property : store.get(ModelKeys.PROPERTY).asPropertyList()) {
                  ModelNode propertyAddress = address.clone().add(ModelKeys.PROPERTY, property.getName());
                  AttributeDefinition[] ATTRIBUTE = {StorePropertyResource.VALUE} ;
-                 result.add(CacheConfigOperationHandlers.createOperation(ATTRIBUTE, propertyAddress, property.getValue()));
+                 result.add(createOperation(ATTRIBUTE, propertyAddress, property.getValue()));
              }
         }
     }
 
-    /*
-    static final DescriptionProvider SUBSYSTEM_DESCRIBE = new DescriptionProvider() {
-        @Override
-        public ModelNode getModelDescription(Locale locale) {
-            return getSubsystemDescribeDescription(locale);
+    static ModelNode createOperation(AttributeDefinition[] attributes, ModelNode address, ModelNode existing) throws OperationFailedException {
+        ModelNode operation = Util.getEmptyOperation(ADD, address);
+        for(final AttributeDefinition attribute : attributes) {
+            attribute.validateAndSet(existing, operation);
         }
-    };
-
-    static ModelNode getSubsystemDescribeDescription(Locale locale) {
-        ResourceBundle resources = getResources(locale);
-        final ModelNode op = createOperationDescription(ModelDescriptionConstants.DESCRIBE, resources, "infinispan.describe");
-        op.get(ModelDescriptionConstants.REQUEST_PROPERTIES).setEmptyObject();
-        op.get(ModelDescriptionConstants.REPLY_PROPERTIES, ModelDescriptionConstants.TYPE).set(ModelType.LIST);
-        op.get(ModelDescriptionConstants.REPLY_PROPERTIES, ModelDescriptionConstants.VALUE_TYPE).set(ModelType.OBJECT);
-        return op;
-    }
-    private static ModelNode createOperationDescription(String operation, ResourceBundle resources, String key) {
-        ModelNode description = new ModelNode();
-        if (operation != null) {
-            description.get(ModelDescriptionConstants.OPERATION_NAME).set(operation);
-        }
-        description.get(ModelDescriptionConstants.DESCRIPTION).set(resources.getString(key));
-        return description;
+        return operation;
     }
 
-    private static ResourceBundle getResources(Locale locale) {
-        return ResourceBundle.getBundle(InfinispanExtension.RESOURCE_NAME, (locale == null) ? Locale.getDefault() : locale);
+    static ModelNode createStoreOperation(AttributeDefinition[] commonAttributes, ModelNode address, ModelNode existing, AttributeDefinition... additionalAttributes) throws OperationFailedException {
+        ModelNode operation = Util.getEmptyOperation(ADD, address);
+        for(final AttributeDefinition attribute : commonAttributes) {
+            attribute.validateAndSet(existing, operation);
+        }
+        for(final AttributeDefinition attribute : additionalAttributes) {
+            attribute.validateAndSet(existing, operation);
+        }
+        return operation;
     }
-    */
+
+    static ModelNode createStringKeyedStoreOperation(ModelNode address, ModelNode existing) throws OperationFailedException {
+        ModelNode operation = Util.getEmptyOperation(ADD, address);
+        for(final AttributeDefinition attribute : COMMON_STORE_ATTRIBUTES) {
+            attribute.validateAndSet(existing, operation);
+        }
+        for(final AttributeDefinition attribute : COMMON_JDBC_STORE_ATTRIBUTES) {
+            attribute.validateAndSet(existing, operation);
+        }
+        for(final AttributeDefinition attribute : STRING_KEYED_JDBC_STORE_ATTRIBUTES) {
+            attribute.validateAndSet(existing, operation);
+        }
+        return operation;
+    }
+
+    static ModelNode createBinaryKeyedStoreOperation(ModelNode address, ModelNode existing) throws OperationFailedException {
+        ModelNode operation = Util.getEmptyOperation(ADD, address);
+        for(final AttributeDefinition attribute : COMMON_STORE_ATTRIBUTES) {
+            attribute.validateAndSet(existing, operation);
+        }
+        for(final AttributeDefinition attribute : COMMON_JDBC_STORE_ATTRIBUTES) {
+            attribute.validateAndSet(existing, operation);
+        }
+        for(final AttributeDefinition attribute : BINARY_KEYED_JDBC_STORE_ATTRIBUTES) {
+            attribute.validateAndSet(existing, operation);
+        }
+        return operation;
+    }
+
+    static ModelNode createMixedKeyedStoreOperation(ModelNode address, ModelNode existing) throws OperationFailedException {
+        ModelNode operation = Util.getEmptyOperation(ADD, address);
+        for(final AttributeDefinition attribute : COMMON_STORE_ATTRIBUTES) {
+            attribute.validateAndSet(existing, operation);
+        }
+        for(final AttributeDefinition attribute : COMMON_JDBC_STORE_ATTRIBUTES) {
+            attribute.validateAndSet(existing, operation);
+        }
+        for(final AttributeDefinition attribute : MIXED_KEYED_JDBC_STORE_ATTRIBUTES) {
+            attribute.validateAndSet(existing, operation);
+        }
+        return operation;
+    }
 
     /*
     * Description provider for the subsystem describe handler
